[Change] separate giftshop controllers from routes
