+++ controllers/info_controllers.js
@@ -1,0 +1,5 @@
+module.exports.findAll;
+module.exports.add;
+module.exports.findById;
+module.exports.editById;
+module.exports.deleteById;
